requirements.txt
pandas>=1.4.0
numpy>=1.21.0
fastapi>=0.68.0
pyspark>=3.2.0
psycopg2-binary>=2.9.3
python-dotenv>=0.19.0
uvicorn>=0.15.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
requests>=2.26.0
aiofiles>=0.8.0
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.12b0
flake8>=4.0.1
mypy>=0.910
uvicorn
asyncpg
slowapi